1/14/24: initial commit
I should've initialized this repo as a git repo earlier, but forgot to do
that earlier in development, so currently everything is in one commit.
Before I opted to create a Rails app to test this, but since we are
just looking at a simple PORO and its corresponding spec, I thought
making a very simple collection of files like this would suffice.