1/14/24: add another failing test
This is another simple test that would emulate an error being raised
if the create! method raised an exception. However, I realize now
that I should've created this as a Rails app, so that I could
explicitly do something like "allow(Log).to receive(:create!).and_raise("error").
However, for the sake of time, I'll leave this test as is, and am
happy to go over it in a pairing session if needed.